WIP - Move camera functions to own module
